dev-requirements.txt
# Requirements that can't be expressed in setup.py
-c https://raw.githubusercontent.com/apache/airflow/constraints-3.0.1/constraints-3.11.txt
wheel==0.45.1
apache-airflow==3.0.1
time_machine
pytest-env

-e .[test]
